+++ xarray/backends/rasterio_.py
@@ -162,7 +162,15 @@ def _parse_envi(meta):
     return parsed_meta
 
 
-def open_rasterio(filename, parse_coordinates=None, chunks=None, cache=None, lock=None, kwargs=None):
+def open_rasterio(
+    filename,
+    parse_coordinates=None,
+    chunks=None,
+    cache=None,
+    lock=None,
+    backend_kwargs=None,
+    **kwargs,
+):
     """Open a file with rasterio (experimental).
 
     This should work with any file that rasterio can open (most often:
@@ -272,7 +280,14 @@ def open_rasterio(filename, parse_coordinates=None, chunks=None, cache=None, loc
     if lock is None:
         lock = RASTERIO_LOCK
 
-    manager = CachingFileManager(rasterio.open, filename, lock=lock, mode="r", kwargs=kwargs)
+    manager = CachingFileManager(
+        rasterio.open,
+        filename,
+        lock=lock,
+        mode="r",
+        backend_kwargs=backend_kwargs,
+        **kwargs,
+    )
     riods = manager.acquire()
     if vrt_params is not None:
         riods = WarpedVRT(riods, **vrt_params)